fixes spec so it passes when the index page works
i had my cart working (displaying all items in the cart including new ones that were added) but this spec wasn't passing.  Changing the spec like above allows for the test to pass if the solution renders the contents of the cart within the index.html.erb view and not by directly assigning the cart in the index action of the ProductsController.